+++ Staging/IFS/Tables/INVOICE_ITEM.sql
@@ -52,7 +52,7 @@
     [N6]                       INT             NULL,
     [N7]                       INT             NULL,
     [N8]                       INT             NULL,
-    [N9]                       INT             NULL,
+    [N9]                       BIGINT          NULL,
     [N10]                      INT             NULL,
     [N11]                      INT             NULL,
     [N12]                      INT             NULL,
@@ -86,4 +86,6 @@
     [BatchID]                  BIGINT          NULL,
     [ProcessLogID]             BIGINT          NULL
 );
+
+
