+++ StagingIFS/PROD/Tables/INVOICE.sql
@@ -74,8 +74,8 @@
     [C12]                    NVARCHAR (100)  NULL,
     [C13]                    NVARCHAR (100)  NULL,
     [N1]                     INT             NULL,
-    [N2]                     INT             NULL,
-    [N3]                     INT             NULL,
+    [N2]                     BIGINT          NULL,
+    [N3]                     BIGINT          NULL,
     [N4]                     INT             NULL,
     [N5]                     INT             NULL,
     [D1]                     DATETIME2 (7)   NULL,
@@ -142,4 +142,6 @@
     [ProcessLogID]           BIGINT          NULL,
     CONSTRAINT [PK_INVOICE] PRIMARY KEY CLUSTERED ([INVOICE_ID] ASC, [COMPANY] ASC)
 );
+
+
